requirements.txt
# ── Django core
Django>=5.0
gunicorn>=21.0

# ── Database
psycopg2-binary>=2.9

# ── Config
python-decouple>=3.8

# ── Media
Pillow>=10.0

# ── Email
resend>=2.0

# ── RAG: Embeddings
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0

# ── RAG: Document loaders
PyPDF2>=3.0.0

# ── Async task queue
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0

# ── Celery result backend dependencies
kombu>=5.3.0
django-celery-beat>=2.0